backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "gradio>=5.50.0",
    "openai>=2.8.1",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
